+++ purescript_waterslide_tests/tests/to_purs_constructor_impls.rs
@@ -1,4 +1,8 @@
 extern crate purescript_waterslide;
+extern crate chrono;
+extern crate uuid;
+
+use chrono::*;
 
 use purescript_waterslide::AsPursConstructor;
 
@@ -27,4 +31,40 @@ fn vecs_derive_as_expected() {
 fn boxes_derive_as_expected() {
     assert_eq!(&format!("{}", <Box<u8>>::as_purs_constructor()), "Int");
 
+}
+
+#[test]
+fn uuid_translates_as_expected() {
+    assert_eq!(&format!("{}", uuid::Uuid::as_purs_constructor()), "String");
+}
+
+#[test]
+fn chrono_datetime_translates_as_expected() {
+    assert_eq!(
+        &format!("{}", DateTime::<Utc>::as_purs_constructor()),
+        "String"
+    );
+}
+
+#[test]
+fn chrono_date_translates_as_expected() {
+    assert_eq!(&format!("{}", Date::<Utc>::as_purs_constructor()), "String");
+}
+
+#[test]
+fn chrono_naive_datetime_translates_as_expected() {
+    assert_eq!(
+        &format!("{}", NaiveDateTime::as_purs_constructor()),
+        "String"
+    );
+}
+
+#[test]
+fn chrono_naive_date_translates_as_expected() {
+    assert_eq!(&format!("{}", NaiveDate::as_purs_constructor()), "String");
+}
+
+#[test]
+fn chrono_naive_time_translates_as_expected() {
+    assert_eq!(&format!("{}", NaiveTime::as_purs_constructor()), "String");
 }