+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+## [0.3.1] 2017-08-08
+
+### Added
+
+- Optional support for `chrono` types and `uuid::Uuid`.
+
 ## [0.3.0] - 2017-08-06
 
 ### Added